packaging: Improve support for downloading from GitHub release
Switch to pep440-pre style
--------------------------

This commit updates the convention used by versioneer
to pep440-pre [1]. Doing so allows to generate packages that do NOT
include a "plus" sign in their names. This is required to support
uploading the prereleases as GitHub release assets so that that can
be installed using pip:

  pip install pynwb --find-links https://github.com/NeurodataWithoutBorders/pynwb/releases/tag/latest  --no-index

Otherwise, GitHub is replacing the "plus" sign with a "dot" and the command
using "--find-links" is failing to find the packages.

[1] https://github.com/warner/python-versioneer/blob/master/details.md#how-do-i-select-a-version-style

Update of scikit-ci-addons
--------------------------

To support upload of package generated using "pep440-pre" convention,
this commit updates the version of scikit-ci-addons to add support
for <COMMIT_DISTANCE> in the mini-language for package selection.

Ensure "latest" is not considered by "versionneer"
--------------------------------------------------

This commit customizes versionner so that it supports "tag_prefix"
specified as glob pattern.

Doing so allows to specify a glob pattern like "*.*.*" ensuring that
only tag matching our convention are considered.

Ensure that latest release date is current
------------------------------------------

The update of scikit-ci-addons addresses this. Fixes #222

Co-authored-by: Doruk Ozturk <doruk.ozturk@kitware.com>