NWBContainer: Remove source attribute
The source attribute is currently a free text entry designed for
provenance attribution.

Experienced showed that this was either mostly ignored or heavily
misused, e.g. sneaking in additional metadata.

Potential readers also don't benefit from the free text approach as it
is difficult to understand the contents without any rules on how to
interpret it.

So let's remove it.

All current users of source can either switch to description or
comments for NWBContainer.