NWBHDF5IO/HDF5IO: Use no default opening mode
The default opening mode for HDF5 files is, in accordance with h5py,
"a" which means "Read/write if exists, create otherwise (default)".

This is for numerous reasons a suboptimal choice:
- On windows this locks files which are maybe supposed to be opened read
  only. Thus other programs can not read these files.
- It opens the potential for data corruption if a file is meant to be
  read only and h5py and its C core crashes.
- It results in rather uninspiring error messages when
  trying to open a non-existing file for reading using

  with NWBHDF5IO("I_DONT_EXIST.nwb") as io:
          io.read()

  -> 5 pages of backtraces following

So let's use no mode as default and force the user to choose if she likes
to read or write.