__builddoc: Create sphinx style links for clickable links
The function __builddoc is responsible for creating the python docstrings
from the docval contents. This docstrings are then converted using
sphinx.ext.napoleon to a format sphinx understands.

On that route it turns out that only links to basic python types are
linked to their respective documentation source.

This is pretty ugly to browse.

We fix that by creating a sphinx style references for known types.

The actual object we currently create the docstring for can not be linked
with this approach. And also NWBContainer does not work.